+++ modules/cells/src/main/java/dmg/cells/services/TelnetLoginManager.java
@@ -58,7 +58,7 @@ public class      TelnetLoginManager
               throw new IllegalArgumentException("USAGE : ... " + __usage);
           }
 
-          _listenPort = new Integer( _args.argv(0) ).intValue() ;
+          _listenPort = new Integer(_args.argv(0));
 
           if( _args.argc() > 1 ) {
               _loginCellClass = _args.argv(1);
