+++ modules/cells/src/main/java/dmg/protocols/snmp/SnmpTest.java
@@ -21,7 +21,7 @@ public class SnmpTest {
      }
    }else{
      String hostName = args[0] ;
-     int port = new Integer( args[1] ).intValue() ;
+     int port = new Integer(args[1]);
      String oid ;
      if( args.length < 3 ) {
          oid = "1.3.6.1.2.1.1.3";
